src/components/not-found-page/not-found-page.test.tsx
import { render, screen } from '@testing-library/react';
import { createMemoryHistory } from 'history';
import HistoryRouter from '../history-route/history-route';
import NotFoundPage from './not-found-page';

describe('Компонент: "NotFoundPage"', () => {
  it('корректно отрисовывается компонент', () => {
    const history = createMemoryHistory();

    render(
      <HistoryRouter history={history}>
        <NotFoundPage />
      </HistoryRouter>,
    );

    expect(screen.getByText('404 :( Такой страницы не существует')).toBeInTheDocument();
    expect(screen.getByText('Главная')).toBeInTheDocument();
    expect(screen.getByText('Несуществующая страница')).toBeInTheDocument();
  });
});
